+++ src/item.h
@@ -147,9 +147,9 @@ public:
  item const* inspect_active_gunmod() const;
  bool goes_bad();
  bool count_by_charges() const;
- int max_charges() const;
+ long max_charges() const;
  bool craft_has_charges();
- int num_charges();
+ long num_charges();
  bool rotten();
  bool ready_to_revive(); // used for corpses
 // light emission, determined by type->light_emission (LIGHT_???) tag (circular),
@@ -217,7 +217,7 @@ public:
 
  std::string name;
  char invlet;           // Inventory letter
- int charges;
+ long charges;
  bool active;           // If true, it has active effects to be processed
  int fridge;            // The turn we entered a fridge.
  int rot;               // decay; same as turn-bday at 65 degrees, but doubles/halves every 18 degrees. can be negative (start game fridges)
